pyproject.toml
[project]
name = "aegis-edr"
version = "0.2.0"
description = "Standalone Endpoint Detection & Response lab"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.3",
    "rich>=15.0.0",
    "typer>=0.24.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "websockets>=13.0",
    "pytest>=8.0.0",
    "watchdog>=4.0.0",
    "httpx>=0.27.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
